Move tests to their corresponding modules.
