+++ core/src/mantle/ops/channel/verification.rs
@@ -50,3 +50,35 @@ pub fn verify_channel_multi_sig(
 
     Ok(())
 }
+
+#[cfg(test)]
+pub mod test_utils {
+    use lb_key_management_system_keys::keys::Ed25519Key;
+
+    use crate::{
+        mantle::{TxHash, ops::channel::ChannelKeyIndex},
+        proofs::channel_multi_sig_proof::{
+            ChannelMultiSigProof, IndexedSignature, IndexedSignatures,
+        },
+    };
+
+    #[must_use]
+    pub fn create_channel_multi_sig_proof(
+        tx_hash: &TxHash,
+        signing_keys: &[&Ed25519Key],
+    ) -> ChannelMultiSigProof {
+        let signatures: IndexedSignatures = signing_keys
+            .iter()
+            .enumerate()
+            .map(|(index, key)| {
+                IndexedSignature::new(
+                    index as ChannelKeyIndex,
+                    key.sign_payload(tx_hash.as_signing_bytes().as_ref()),
+                )
+            })
+            .collect::<Vec<_>>()
+            .try_into()
+            .unwrap();
+        ChannelMultiSigProof::try_new(signatures).unwrap()
+    }
+}
